requirements.txt
Flask
Flask-MySQL
python-dotenv